requirements.txt
beautifulsoup4
langchain
langchain-community
langchain-openai
langchain-chroma
python-dotenv
streamlit
chromadb
requests
